Avoid using Str
The "str" library is still shipped together with the OCaml stdlib but generally
considered obsolete; better alternatives include the `re` library that we use in
Catala, but here, the use-case being very simple, `Scanf` is powerful enough and
reasults in more readable code.

To be honest, my main motivation was avoiding the need to link another
dependency into the Catala runtime :)

Note/advice: side effects within `assert` should be avoided, `assert` in OCaml
is not a standard function and it would break your code if compiled with
the `-noassert` flag.